refactor auth handlers to context

+++ client/src/context/Context.js
@@ -1,29 +1,34 @@
 import React, { createContext, useEffect, useState } from 'react';
-import axios from 'axios';
+import * as api from '../api';
 
 export const myContext = createContext({});
 
 export default function Context(props) {
   const [user, setUser] = useState(localStorage.getItem('user'));
 
-  useEffect(() => {
-    axios
-      .get('http://localhost:5000/user', { withCredentials: true })
-      .then((res) => {
-        if (res.data) {
-          localStorage.setItem('user', res.data);
-          setUser(localStorage.getItem('user'));
-        }
-      });
-  }, []);
-
-  const clearUser = () => {
-    localStorage.clear();
-    // setUser(null);
+  const handleLogout = () => {
+    api.googleLogout().then((res) => {
+      if (res.data === 'done') {
+        localStorage.clear();
+        window.location.href = 'http://localhost:3000/login';
+      }
+    });
   };
 
+  const handleLogin = () => {
+    window.location.href = 'http://localhost:5000/auth/google';
+  };
+
+  useEffect(() => {
+    api.fetchUser().then((res) => {
+      if (res.data) {
+        localStorage.setItem('user', res.data);
+      }
+    });
+  }, []);
+
   return (
-    <myContext.Provider value={{ user, clearUser }}>
+    <myContext.Provider value={{ user, handleLogin, handleLogout }}>
       {props.children}
     </myContext.Provider>
   );
